Float support

+++ include/programfile.h
@@ -6,7 +6,7 @@ class BinaryProgram {
 public:
     std::vector<uint8_t> bytecode;
     std::vector<std::string> stringPool;
-    std::vector<int> constPool;
+    std::vector<double> constPool;
     int variableCount;
     bool save(const std::string& path);
     bool load(const std::string& path);
